--- a/app/views/contabilidad/reportes/cash_flow.html.slim
+++ b/app/views/contabilidad/reportes/cash_flow.html.slim
@@ -11,169 +11,40 @@ h1.pull-left Flujo de Efectivo
 
 br
 
-- date_range = @from_date..@from_date.end_of_year
-- date_months = date_range.map { |d| Date.new(d.year, d.month, 1) }.uniq
-
-- initials = []
-- revenues = []
-- expenses = []
-- liabilities = []
-- equities = []
-- assets = []
-
 table.cash-flow
   thead.thead-dark
     tr.text-center
       th scope="col"=@from_date.to_s.first(4)
-      - date_months.each do |monthly|
+      - @date_months.each do |monthly|
         th scope="col"
           =link_to I18n.t('date.month_names')[monthly.month], contabilidad_entradas_path(limit: 300, order: 'descendente', start: monthly, end: monthly.end_of_month)
 
   tbody
-    tr
-      th scope="row"
-        span> class="#{plutus_to_badge('Plutus::Asset')}" &nbsp;
-        | Saldo Inicial del Periodo
-
-    - @initials.each do |account|   
-      tr
-        td
-          em=account.name
-        - date_months.each do |monthly|
-          - balance = account.balance(from_date: Cuenta::EPOCH, to_date: monthly - 1.day)
-          td
-            em=link_to_account balance, monthly - 1.month, account
-
-    tr
-      th.text-right scope="row" Total de Saldo Inicial
-      - date_months.each do |monthly|
-        - initials << Cuenta.caja_chica.balance(from_date: Cuenta::EPOCH, to_date: monthly - 1.day) + Cuenta.caja_fuerte.balance(from_date: Cuenta::EPOCH, to_date: monthly - 1.day) + Cuenta.banco.balance(from_date: Cuenta::EPOCH, to_date: monthly - 1.day)
-        td
-          strong=initials.last
-    
-    tr
-      th scope="row"
-        span> class="#{plutus_to_badge('Plutus::Revenue')}" &nbsp;
-        | Ingresos
-    
-    - @revenues.each do |account|   
-      tr
-        td
-          em=account.name
-        - date_months.each do |monthly|
-          - balance = account.balance(from_date: monthly, to_date: monthly.end_of_month)
-          td
-            em=link_to_account balance, monthly, account
-
-    tr
-      th.text-right scope="row" Total de Ingresos
-      - date_months.each do |monthly|
-        - revenues << Plutus::Revenue.balance(from_date: monthly, to_date: monthly.end_of_month)
-        td
-          strong=revenues.last
-
-    tr
-      th scope="row"
-        span> class="#{plutus_to_badge('Plutus::Expense')}" &nbsp;
-        | Egresos
-
-    - @expenses.each do |account|   
-      tr
-        td
-          em=account.name
-        - date_months.each do |monthly|
-          - balance = account.balance(from_date: monthly, to_date: monthly.end_of_month)
-          td
-            em=link_to_account balance, monthly, account
-
-    tr
-      th.text-right scope="row" Total de Egresos
-      - date_months.each do |monthly|
-        - expenses << Plutus::Expense.balance(from_date: monthly, to_date: monthly.end_of_month)
-        td
-          strong=expenses.last
-
-  tbody
-    tr
-      th scope="row"
-        span> class="#{plutus_to_badge('Plutus::Asset')}" &nbsp;
-        | Inversiones
-
-    - @assets.each do |account|   
-      tr
-        td
-          em=account.name
-        - date_months.each do |monthly|
-          - balance = account.balance(from_date: monthly, to_date: monthly.end_of_month)
-          td
-            em=link_to_account balance, monthly, account
-
-    tr
-      th.text-right scope="row" Total de Inversiones
-      - date_months.each do |monthly|
-        - assets << Plutus::Asset.balance(from_date: monthly, to_date: monthly.end_of_month) - @initials.map { |initial| initial.balance(from_date: monthly, to_date: monthly.end_of_month) }.sum
-        td
-          strong=assets.last
+    =render partial: 'cash_flow_accounts', locals: { name: 'Saldo Inicial', data: @initials }
+    =render partial: 'cash_flow_accounts', locals: { name: 'Ingresos', data: @revenues }
+    =render partial: 'cash_flow_accounts', locals: { name: 'Egresos', data: @expenses }
+    =render partial: 'cash_flow_accounts', locals: { name: 'Activos', data: @assets }
 
   tbody.tfoot
     tr
       th scope="row"
         span> class="#{plutus_to_badge('Plutus::Asset')}" &nbsp;
         | Resultado de la Actividad
-      - date_months.each_with_index do |monthly, index|
+      - @date_months.each_with_index do |monthly, index|
         td
-          strong=initials[index] + revenues[index] - expenses[index] - assets[index]
+          strong=@initials[:totals][index] + @revenues[:totals][index] - @expenses[:totals][index] - @assets[:totals][index]
     tr
       td
 
   tbody
-    tr
-      th scope="row"
-        span> class="#{plutus_to_badge('Plutus::Liability')}" &nbsp;
-        | Financiamiento
-
-    - @liabilities.each do |account|   
-      tr
-        td
-          em=account.name
-        - date_months.each do |monthly|
-          - balance = account.balance(from_date: monthly, to_date: monthly.end_of_month)
-          td
-            em=link_to_account balance, monthly, account
-
-    tr
-      th.text-right scope="row" Total de Financiamiento
-      - date_months.each do |monthly|
-        - liabilities << Plutus::Liability.balance(from_date: monthly, to_date: monthly.end_of_month)
-        td
-          strong=liabilities.last
-
-    tr
-      th scope="row"
-        span> class="#{plutus_to_badge('Plutus::Equity')}" &nbsp;
-        | Aportaciones
-
-    - @equities.each do |account|   
-      tr
-        td
-          em=account.name
-        - date_months.each do |monthly|
-          - balance = account.balance(from_date: monthly, to_date: monthly.end_of_month)
-          td
-            em=link_to_account balance, monthly, account
-
-    tr
-      th.text-right scope="row" Total de Aportaciones
-      - date_months.each do |monthly|
-        - equities << Plutus::Equity.balance(from_date: monthly, to_date: monthly.end_of_month)
-        td
-          strong=equities.last
+    =render partial: 'cash_flow_accounts', locals: { name: 'Financiamiento', data: @liabilities }
+    =render partial: 'cash_flow_accounts', locals: { name: 'Aportaciones', data: @equities }
 
   tfoot
     tr
       th scope="row"
         span> class="#{plutus_to_badge('Plutus::Asset')}" &nbsp;
         | Total Flujo de Caja Financiero 
-      - date_months.each_with_index do |monthly, index|
+      - @date_months.each_with_index do |monthly, index|
         td
-          strong=initials[index] + revenues[index] + equities[index] - expenses[index] - assets[index] + liabilities[index]
+          strong=@initials[:totals][index] + @revenues[:totals][index] + @equities[:totals][index] - @expenses[:totals][index] - @assets[:totals][index] + @liabilities[:totals][index]