old:
Completed 200 OK in 3947ms (Views: 3351.3ms | ActiveRecord: 528.1ms | Allocations: 2443810)

new:
Completed 200 OK in 1564ms (Views: 50.6ms | ActiveRecord: 468.5ms | Allocations: 999253)